// src/app/models/registracijastudent.ts
export class RegistrovaniClan {
    ime: String;
    prezime: String;
    email: String;
    emailFakultet: String;
    brojTelefona: String;
    brojIndeksa: String;
    korisnickoIme: String;
    lozinka: String;
    odsjek: String;
    godinaStudija: String;
    constructor(ime: String, prezime: String, email: String, emailFakultet: String, brojTelefona: String, brojIndeksa: String,
    korisnickoIme: String, lozinka: String, odsjek: String, godinaStudija: String) {
        this.ime=ime;
        this.prezime=prezime;
        this.email=email;
        this.emailFakultet=emailFakultet;
        this.brojTelefona=brojTelefona;
        this.brojIndeksa=brojIndeksa;
        this.korisnickoIme=korisnickoIme;
        this.lozinka=lozinka;
        this.odsjek=odsjek;
        this.godinaStudija=godinaStudija;
    }
    }